# ARM Neon optimization of `oj_dump_cstr`