@samyron 2025-05-11
The performance is pretty consistent after the simplifications.

```
== Encoding activitypub.json (52595 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after     2.230k i/100ms
Calculating -------------------------------------
               after     22.419k (± 1.6%) i/s   (44.60 μs/i) -    113.730k in   5.074151s

Comparison:
              before:    17756.7 i/s
               after:    22419.4 i/s - 1.26x  faster


== Encoding citm_catalog.json (500298 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after    90.000 i/100ms
Calculating -------------------------------------
               after    905.970 (± 1.5%) i/s    (1.10 ms/i) -      4.590k in   5.067532s

Comparison:
              before:      916.5 i/s
               after:      906.0 i/s - same-ish: difference falls within error


== Encoding twitter.json (466906 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after   220.000 i/100ms
Calculating -------------------------------------
               after      2.204k (± 1.3%) i/s  (453.63 μs/i) -     11.220k in   5.090643s

Comparison:
              before:     1999.9 i/s
               after:     2204.4 i/s - 1.10x  faster
```